tracker: resolve agent contracts server-side
